requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.8.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
autopage==0.5.1
Babel==2.11.0
backcall==0.2.0
basemap==1.3.6
basemap-data==1.3.2
beautifulsoup4==4.11.1
black==22.10.0
bleach==5.0.1
cachetools==5.2.0
certifi==2023.7.22
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.3
click-plugins==1.1.1
cliff==4.1.0
cligj==0.7.2
cmaes==0.9.0
cmake==3.27.2
cmd2==2.4.2
colorlog==6.7.0
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
docker-pycreds==0.4.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.12.2
Fiona==1.9.3
flake8==6.0.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.2
geopandas==0.12.2
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.15.0
google-auth-oauthlib==0.4.6
greenlet==2.0.1
grpcio==1.51.1
h5py==3.7.0
hydra-colorlog==1.2.0
hydra-core==1.2.0
hydra-optuna-sweeper==1.2.0
identify==2.5.9
idna==3.4
importlib-metadata==5.1.0
iniconfig==1.1.1
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
isort==5.10.1
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonschema==4.17.3
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyter-server==1.23.3
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.3
kaleido==0.2.1
kiwisolver==1.4.4
lit==16.0.6
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.3
matplotlib==3.6.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
mpmath==1.3.0
multidict==6.0.2
munch==2.5.0
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.5
nbformat==5.7.0
nbstripout==0.6.1
nest-asyncio==1.5.5
networkx==3.1
nodeenv==1.7.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.2.3
openpyxl==3.0.10
optuna==2.10.1
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.2
pathtools==0.1.2
pbr==5.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pip==22.2.2
platformdirs==2.5.4
plotly==5.13.1
plotly-geo==1.0.0
pluggy==1.0.0
pre-commit==2.20.0
prettytable==3.5.0
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pudb==2022.1.3
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyDeprecate==0.3.1
pyflakes==3.0.1
Pygments==2.13.0
pyparsing==3.0.9
pyperclip==1.8.2
pyproj==3.4.1
pyrootutils==1.0.4
pyrsistent==0.19.2
pyshp==2.3.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytorch-lightning==1.5.10
pytz==2022.6
PyYAML==6.0
pyzmq==23.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==12.4.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.0
sentry-sdk==1.11.1
setproctitle==1.3.2
setuptools==68.1.2
sh==1.14.2
shapely==2.0.1
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
stack-data==0.2.0
stevedore==4.1.1
sympy==1.12
tenacity==8.2.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.0
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchmetrics==0.8.2
torchvision==0.15.2+cu117
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
triton==2.0.0
typing_extensions==4.7.1
urllib3==2.0.4
urwid==2.1.2
urwid-readline==0.13
virtualenv==20.17.0
wandb==0.13.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
wheel==0.41.1
xlrd==2.0.1
yarl==1.8.1
zipp==3.11.0
